requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-nvidia-ai-endpoints>=0.1.0
langchain-anthropic>=0.1.0
elasticsearch[async]>=8.13.0
pydantic-settings>=2.0.0
apscheduler>=3.10.0
httpx>=0.27.0
python-dotenv>=1.0.0
redis>=5.0.0
